requirements.txt
Django==2.2
psycopg2-binary==2.8.1
djangorestframework==3.9.2
django-filter==2.1.0
coreapi==2.3.3
PyYAML==5.1
pygments
markdown